# packages/types/openapi.yaml
openapi: 3.1.0
info:
  title: Mirage API
  version: 0.0.0
  description: |
    The single OpenAPI spec for Mirage. Lives in `@mirage/types` so it can
    be consumed both by the BFF (for route shape) and the React SPA (for the
    typed client). Generated TypeScript types land in `openapi.generated.ts`
    via `pnpm gen:openapi`, which is wired to `postinstall`.
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
paths:
...
  /workspaces:
    get:
      summary: List the current org's Workspaces
      operationId: listWorkspaces
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspace'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a Workspace
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /workspaces/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      summary: Get a Workspace by id
      operationId: getWorkspace
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a Workspace (cascade)
      description: |
        Soft-deletes the workspace immediately (sets `deletedAt`) and starts
        an async cascade that cancels in-flight runs, purges every S3 object
        under `org/{orgId}/workspace/{wsId}/`, and deletes every Mongo doc
        scoped to this workspace. Caller must be an org owner.
      operationId: deleteWorkspace
      responses:
        '202':
          description: Accepted — soft delete written; cascade in progress
          content:
            application/json:
              schema:
                type: object
                required: [id, deletedAt]
                additionalProperties: false
                properties:
                  id: { type: string }
                  deletedAt: { type: string, format: date-time }
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Workspace is already being deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /workspaces/{wsId}/custom-functions:
    parameters:
      - in: path
        name: wsId
        required: true
        schema: { type: string }
    get:
      summary: List Custom Functions in a Workspace
      operationId: listCustomFunctions
      parameters:
        - in: query
          name: usage
          required: false
          schema: { type: string, enum: [valueGenerator, strategy, both] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CustomFunction' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      summary: Create a Custom Function
      operationId: createCustomFunction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCustomFunctionBody' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CustomFunction' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /workspaces/{wsId}/custom-functions/{id}:
    parameters:
      - in: path
        name: wsId
        required: true
        schema: { type: string }
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      summary: Get a Custom Function by id
      operationId: getCustomFunction
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CustomFunction' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Update a Custom Function
      operationId: updateCustomFunction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCustomFunctionBody' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CustomFunction' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      summary: Delete a Custom Function
      operationId: deleteCustomFunction
      responses:
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Workspace:
      type: object
      required: [id, orgId, name, createdBy, createdAt, updatedAt]
      additionalProperties: false
      properties:
        id: { type: string }
        orgId: { type: string }
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500 }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt:
          type: string
          format: date-time
          description: |
            Set when the workspace is soft-deleted. While present, the row is
            still returned by list/get but every sub-resource endpoint
            responds 404. The row is hard-deleted (removed entirely) once the
            background cascade finishes.
    CreateWorkspaceBody:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 500 }
...
    CustomFunction:
      type: object
      required:
        - id
        - workspaceId
        - orgId
        - name
        - description
        - usage
        - source
        - createdBy
        - createdAt
        - updatedAt
      additionalProperties: false
      properties:
        id: { type: string }
        workspaceId: { type: string }
        orgId: { type: string }
        name:
          type: string
          pattern: '^[a-zA-Z_$][a-zA-Z0-9_$]{0,63}$'
        description: { type: string, maxLength: 500 }
        usage: { type: string, enum: [valueGenerator, strategy, both] }
        source: { type: string, minLength: 1, maxLength: 20000 }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CreateCustomFunctionBody:
      type: object
      required: [name, usage, source]
      additionalProperties: false
      properties:
        name:
          type: string
          pattern: '^[a-zA-Z_$][a-zA-Z0-9_$]{0,63}$'
        description: { type: string, maxLength: 500 }
        usage: { type: string, enum: [valueGenerator, strategy, both] }
        source: { type: string, minLength: 1, maxLength: 20000 }
    UpdateCustomFunctionBody:
      type: object
      required: [name, usage, source, expectedUpdatedAt]
      additionalProperties: false
      properties:
        name:
          type: string
          pattern: '^[a-zA-Z_$][a-zA-Z0-9_$]{0,63}$'
        description: { type: string, maxLength: 500 }
        usage: { type: string, enum: [valueGenerator, strategy, both] }
        source: { type: string, minLength: 1, maxLength: 20000 }
        expectedUpdatedAt: { type: string, format: date-time }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        detail: {}
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Unauthenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }